requirements.txt
Scrapy~=2.6.2
omegaconf~=2.2.2
tqdm~=4.64.0
hydra-core~=1.2.0